CHANGES: snmpd: PATCH: 3369680: from Ilya Yanok: change snmp_enableauthentraps from int to long such that modifying the corresponding MIB object does not overwrite memory in 64-bit builds.

+++ agent/agent_trap.c
@@ -117,7 +117,7 @@ char           *snmp_trapcommunity = NULL;
 #define SNMP_AUTHENTICATED_TRAPS_ENABLED	1
 #define SNMP_AUTHENTICATED_TRAPS_DISABLED	2
 
-int             snmp_enableauthentraps = SNMP_AUTHENTICATED_TRAPS_DISABLED;
+long            snmp_enableauthentraps = SNMP_AUTHENTICATED_TRAPS_DISABLED;
 int             snmp_enableauthentrapsset = 0;
 
 /*
