+++ agent/mibgroup/mibII/snmp_mib_5_5.c
@@ -12,7 +12,7 @@
 
 static oid snmp_oid[] = { SNMP_OID };
 
-extern int snmp_enableauthentraps;
+extern long snmp_enableauthentraps;
 extern int snmp_enableauthentrapsset;
 
 static int
@@ -21,7 +21,7 @@ snmp_enableauthentraps_store(int a, int b, void *c, void *d)
     char            line[SNMP_MAXBUF_SMALL];
 
     if (snmp_enableauthentrapsset > 0) {
-        snprintf(line, SNMP_MAXBUF_SMALL, "pauthtrapenable %d",
+        snprintf(line, SNMP_MAXBUF_SMALL, "pauthtrapenable %ld",
                  snmp_enableauthentraps);
         snmpd_store_config(line);
     }
